Implemen other methods post participation flow start
